requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pip==23.0.1
python-decouple==3.8
pytz==2022.7.1
setuptools==67.6.0
typing-extensions==4.5.0
wheel==0.38.4
yarl==1.8.2
